feat(blade): add base text for web and native with code sharing
