+++ .eslintrc.js
@@ -86,6 +86,8 @@ module.exports = {
             extensions: ['.ts', '.tsx'],
           },
         ],
+        'babel/new-cap': ['error', { capIsNewExceptionPattern: '^styled.' }],
+        '@typescript-eslint/consistent-type-imports': ['error', { prefer: 'type-imports' }],
         '@typescript-eslint/no-unsafe-assignment': 'off',
         '@typescript-eslint/no-unused-vars': ['error'],
         // '@typescript-eslint/no-unsafe-return': 'off',
