--- packages/blade/src/components/Typography/BaseText.d.ts
@@ -1,1 +1,0 @@
-export { default } from './BaseText.web';
