+++ packages/blade/src/components/Typography/BaseText/StyledBaseText.d.ts
@@ -1,0 +1,1 @@
+export { default } from './StyledBaseText.web';
