+++ packages/blade/src/components/Typography/BaseText/StyledBaseText.native.tsx
@@ -1,0 +1,30 @@
+import styled from 'styled-components/native';
+import getBaseTextStyles from './getBaseTextStyles';
+import type { BaseTextProps } from './BaseText.d';
+
+const StyledBaseText = styled.Text<BaseTextProps>(
+  ({
+    theme,
+    color,
+    fontFamily,
+    fontSize,
+    fontWeight,
+    fontStyle,
+    textDecorationLine,
+    lineHeight,
+    textAlign,
+  }) =>
+    getBaseTextStyles({
+      theme,
+      color,
+      fontFamily,
+      fontSize,
+      fontWeight,
+      fontStyle,
+      textDecorationLine,
+      lineHeight,
+      textAlign,
+    }),
+);
+
+export default StyledBaseText;
