+++ packages/blade/src/components/Typography/BaseText/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BaseText';
